+++ moyu-common/src/main/java/cn/xueden/common/vo/edu/EduChapterTreeNodeVO.java
@@ -34,7 +34,7 @@ public class EduChapterTreeNodeVO {
 
     private int lev;
 
-    private Long duration;
+    private Float duration;
 
     private List<EduChapterTreeNodeVO> children;
