+++ moyu-common/src/main/java/cn/xueden/common/vo/edu/EduChapterVO.java
@@ -23,7 +23,7 @@ public class EduChapterVO {
 
     private String title;
 
-    private Long duration;
+    private Float duration;
 
     private Integer sort;
